+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - configurable configurator set type
 - configurable variation number field
+- configurable order origin (referrer)
 
 ### Fixed
 - optimized the update path for versions before 2.0